chore(nlp): Renames existing modules to their new nomenclatures.
