+++ archai/nlp/models/data_utils.py
@@ -20,8 +20,8 @@ import re
 import sacremoses
 import torch
 
-from archai.nlp.nvidia_transformer_xl import nvidia_utils as nv_utils
-from archai.nlp.nvidia_transformer_xl.nvidia_utils.corpus import Corpus
+from archai.nlp.models import nvidia_utils as nv_utils
+from archai.nlp.models.nvidia_utils.corpus import Corpus
 
 def get_lm_corpus(datadir:str, cachedir:str, dataset:str, vocab_type:str,
                   vocab_size:Optional[int]=None, refresh_cache=False):
