+++ archai/nlp/models/models/mem_transformer.py
@@ -21,12 +21,12 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from archai.nlp.nvidia_transformer_xl.nvidia_utils.log_uniform_sampler import LogUniformSampler
-from archai.nlp.nvidia_transformer_xl.nvidia_utils.log_uniform_sampler import sample_logits
-from archai.nlp.nvidia_transformer_xl.nvidia_utils.proj_adaptive_softmax import ProjectedAdaptiveLogSoftmax
-from archai.nlp.nvidia_transformer_xl.primer_ez import DWiseConvPrimerEZ, PositionwiseFFPrimerEZ
-from archai.nlp.nvidia_transformer_xl.models.archai_model import ArchaiModel
-from archai.nlp.nvidia_transformer_xl.models.model_utils import map_to_list
+from archai.nlp.models.nvidia_utils.log_uniform_sampler import LogUniformSampler
+from archai.nlp.models.nvidia_utils.log_uniform_sampler import sample_logits
+from archai.nlp.models.nvidia_utils.proj_adaptive_softmax import ProjectedAdaptiveLogSoftmax
+from archai.nlp.models.primer_ez import DWiseConvPrimerEZ, PositionwiseFFPrimerEZ
+from archai.nlp.models.models.archai_model import ArchaiModel
+from archai.nlp.models.models.model_utils import map_to_list
 
 
 @torch.jit.script
