+++ archai/nlp/models/onnx/onnx_utils/optimization.py
@@ -8,9 +8,9 @@ from onnx import load_model
 from onnxruntime.transformers.onnx_model_gpt2 import Gpt2OnnxModel
 from onnxruntime.transformers.optimizer import optimize_by_onnxruntime
 
-from archai.nlp.nvidia_transformer_xl.onnx.onnx_utils.load import create_file_name_identifier
-from archai.nlp.nvidia_transformer_xl.onnx.onnx_utils.opt.fusion_options import FusionOptions
-from archai.nlp.nvidia_transformer_xl.onnx.onnx_utils.opt.opt_models import MemTransformerLMOnnxModel
+from archai.nlp.models.onnx.onnx_utils.load import create_file_name_identifier
+from archai.nlp.models.onnx.onnx_utils.opt.fusion_options import FusionOptions
+from archai.nlp.models.onnx.onnx_utils.opt.opt_models import MemTransformerLMOnnxModel
 
 # List of available ONNX models to be optimized
 AVAILABLE_ONNX_OPTS = {
