+++ archai/nlp/models/onnx/onnx_utils/quantization.py
@@ -12,7 +12,7 @@ from onnxruntime.quantization.quant_utils import attribute_to_kwarg, ms_domain
 from onnxruntime.quantization.quantize import quantize_dynamic
 from onnxruntime.quantization.registry import IntegerOpsRegistry
 
-from archai.nlp.nvidia_transformer_xl.onnx.onnx_utils.load import create_file_name_identifier
+from archai.nlp.models.onnx.onnx_utils.load import create_file_name_identifier
 
 
 class GemmQuant(QuantOperatorBase):
