+++ archai/nlp/models/search.py
@@ -14,8 +14,8 @@ import yaml
 from numpy.core.numeric import indices
 
 from archai.common import utils
-from archai.nlp.nvidia_transformer_xl.search_utils.constraint_getter import get_latency, get_model, process_parameters, get_yaml_values
-from archai.nlp.nvidia_transformer_xl.search_utils.info_getter import get_metrics, get_results
+from archai.nlp.models.search_utils.constraint_getter import get_latency, get_model, process_parameters, get_yaml_values
+from archai.nlp.models.search_utils.info_getter import get_metrics, get_results
 
 model_config_defaults = {'d_head': None,
                          'n_token': 267736,
