+++ scripts/nlp/test_score.py
@@ -6,9 +6,9 @@ from typing import Optional
 import torch
 
 from archai.nlp.scoring import score
-from archai.nlp.nvidia_transformer_xl.mem_transformer import MemTransformerLM
-from archai.nlp.nvidia_transformer_xl.data_utils import get_lm_corpus
-from archai.nlp.nvidia_transformer_xl.nvidia_utils import exp_utils
+from archai.nlp.models.mem_transformer import MemTransformerLM
+from archai.nlp.models.data_utils import get_lm_corpus
+from archai.nlp.models.nvidia_utils import exp_utils
 from archai.common import utils, common
 from archai.nlp.tokenizer_utils.vocab_base import VocabBase
